fix/rename entity
