+++ backend/src/chatMuteUser/chatMuteUser.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ChatMuteUserController } from './chatMuteUser.controller';
-import { ChatMuteUser } from '../entity/chatMuteUser.entity';
+import { ChatMuteUser } from '../entities/chatMuteUser.entity';
 import { ChatMuteUserService } from './chatMuteUser.service';
 
 @Module({
