+++ backend/src/entities/chatMuteUser.entity.ts
