+++ backend/src/entities/healthCheck.entity.ts
